bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Telegram Bot with test mode"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]
